backend/requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_jwt_extended
pydantic
python-dotenv
pymysql
